created possiibility to have general CloudTrail processing code placed firehose constants in firehose moved delete_ids to check LookupEvents shoud
